refactor: move AI feed route logic into dedicated controller methods #540
